Rename a few more files
